Agw based ackmode using monitor own.
The way this works is to first enable monitoring on the AGW connection so
that we'll be informed of 'own information' frames as they're transmitted.

This works pretty similar to how XKISS ACKMODE works on DW's side (and
could probably be integrated in a similar way to make *that* patch more
acceptable) Basically after a frame is sent (in DW's send_one_frame) a
summary packet is sent back to *all* AGW clients that have monitoring
enabled.

It's important to notice that all clients receive the monitoring packet,
not just packets their connection sent. The right behavior, but means we
have to be careful.

Next, BPQ maintains a *single* FIFO queue in a circular buffer for the AGW
connection. Each time a frame is sent, an entry is queued and the LINK
associated with that frame is stored, if one is available.
The 'Linkptr' is only there if the frame requires the T2 timer to be set.
It's important to maintain a single queue for the AGW connection because
otherwise correlating the monitored frame to which outgoing frame it
corresponds to is tricky, and we won't know when we should actually
enable the timer. This way, whenever we see a monitored frame, we pop
from the front and if the LINK pointer is set on that queue entry we
can start the timer. By storing a queue entry for every frame, regardless
of how it may affect the T2 timer, we avoid having to inspect the confirmed
frames.

The only other thing we need to do is remember to ignore monitored frames
for origins not associated with any of our links or ourselves. This way
if another AGW client is using the server to send frames we don't get
confused and our queue stays balanced, because as mentioned earlier each
AGW connection sees all monitored frames.