Register calls for all UZ7HO ports on AGW connection.
If multiple ports use the same AGW server then BPQ will multiplex
all those ports on a single TCP connection to that AGW server. When
BPQ goes to register the callsigns that the AGW server should "listen"
for, it only ever registers the callsigns on "AGW Port 0" ignoring
the CHANNEL configuration for that port.

This fix does two things:

First, the channel information is sent for the AGW Port, this is the
most important step. So Channel C maps to Port 2, for example, rather
than always using Port 0.

Second, when the AGW connection is made, we loop over all configured ports
and register those calls for those Channels.